Switch entities

+++ Server/Java/src/main/java/com/maxlvshv/taskflow/repository/ProjectRepo.java
@@ -3,5 +3,5 @@ package com.maxlvshv.taskflow.repository;
 import com.maxlvshv.taskflow.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepo extends JpaRepository<UserEntity, Long> {
+public interface ProjectRepo extends JpaRepository<UserEntity, Integer> {
 }
