+++ Server/Java/src/main/java/com/maxlvshv/taskflow/repository/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<UserEntity, Long> {
+public interface UserRepo extends JpaRepository<UserEntity, Integer> {
     Optional<UserEntity> findByNickname(String nickname);
 }
